refactor(ui): extract lightweight tool renderers to tools-basic module
Slice 11b.1 step 10 of the message-part.tsx split. Move the
BasicTool-wrapper tool renderers — every tool whose render body is a
direct BasicTool with at most a small For/Show body — into one file.

New: packages/ui/src/components/message-part-tools-basic.tsx
  Side-effect registers:
    read · list · glob · grep · webfetch · websearch
    enter-worktree · exit-worktree · todowrite · question · skill

message-part.tsx now imports the module purely as a side-effect; the
larger renderers (task / agent / bash / edit / write / apply_patch) stay
behind for the next round of extraction.

tsgo --noEmit GREEN. bun test: 478 pass / 3 pre-existing fail.
