+++ packages/ui/src/components/message-part.tsx
@@ -202,220 +202,10 @@ export type { ToolProps, ToolComponent }
 import { MessageDivider } from "./message-part-core-renderers"
 export { MessageDivider }
 
-ToolRegistry.register({
-  name: "read",
-  render(props) {
-    const data = useData()
-    const i18n = useI18n()
-    const args: string[] = []
-    if (props.input.offset) args.push("offset=" + props.input.offset)
-    if (props.input.limit) args.push("limit=" + props.input.limit)
-    const loaded = createMemo(() => {
-      if (props.status !== "completed") return []
-      const value = props.metadata.loaded
-      if (!value || !Array.isArray(value)) return []
-      return value.filter((p): p is string => typeof p === "string")
-    })
-    return (
-      <>
-        <BasicTool
-          {...props}
-          icon="glasses"
-          trigger={{
-            title: i18n.t("ui.tool.read"),
-            subtitle: props.input.filePath ? getFilename(props.input.filePath) : "",
-            args,
-          }}
-        />
-        <For each={loaded()}>
-          {(filepath) => (
-            <div data-component="tool-loaded-file">
-              <Icon name="enter" />
-              <span>
-                {i18n.t("ui.tool.loaded")} {relativizeProjectPath(filepath, data.directory)}
-              </span>
-            </div>
-          )}
-        </For>
-      </>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "list",
-  render(props) {
-    const i18n = useI18n()
-    return (
-      <BasicTool
-        {...props}
-        icon="bullet-list"
-        trigger={{ title: i18n.t("ui.tool.list"), subtitle: getDirectory(props.input.path || "/") }}
-      >
-        <Show when={props.output}>
-          <div data-component="tool-output" data-scrollable>
-            <MessageMarkdown text={props.output!} />
-          </div>
-        </Show>
-      </BasicTool>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "glob",
-  render(props) {
-    const i18n = useI18n()
-    return (
-      <BasicTool
-        {...props}
-        icon="magnifying-glass-menu"
-        trigger={{
-          title: i18n.t("ui.tool.glob"),
-          subtitle: getDirectory(props.input.path || "/"),
-          args: props.input.pattern ? ["pattern=" + props.input.pattern] : [],
-        }}
-      >
-        <Show when={props.output}>
-          <div data-component="tool-output" data-scrollable>
-            <MessageMarkdown text={props.output!} />
-          </div>
-        </Show>
-      </BasicTool>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "grep",
-  render(props) {
-    const i18n = useI18n()
-    const args: string[] = []
-    if (props.input.pattern) args.push("pattern=" + props.input.pattern)
-    if (props.input.include) args.push("include=" + props.input.include)
-    return (
-      <BasicTool
-        {...props}
-        icon="magnifying-glass-menu"
-        trigger={{
-          title: i18n.t("ui.tool.grep"),
-          subtitle: getDirectory(props.input.path || "/"),
-          args,
-        }}
-      >
-        <Show when={props.output}>
-          <div data-component="tool-output" data-scrollable>
-            <MessageMarkdown text={props.output!} />
-          </div>
-        </Show>
-      </BasicTool>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "webfetch",
-  render(props) {
-    const i18n = useI18n()
-    const pending = createMemo(() => props.status === "pending" || props.status === "running")
-    const url = createMemo(() => {
-      const value = props.input.url
-      if (typeof value !== "string") return ""
-      return value
-    })
-    return (
-      <BasicTool
-        {...props}
-        hideDetails
-        icon="window-cursor"
-        trigger={
-          <div data-slot="basic-tool-tool-info-structured">
-            <div data-slot="basic-tool-tool-info-main">
-              <span data-slot="basic-tool-tool-title">
-                <TextShimmer text={i18n.t("ui.tool.webfetch")} active={pending()} />
-              </span>
-              <Show when={!pending() && url()}>
-                <a
-                  data-slot="basic-tool-tool-subtitle"
-                  class="clickable subagent-link"
-                  href={url()}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  onClick={(event) => event.stopPropagation()}
-                >
-                  {url()}
-                </a>
-              </Show>
-            </div>
-            <Show when={!pending() && url()}>
-              <div data-component="tool-action">
-                <Icon name="square-arrow-top-right" />
-              </div>
-            </Show>
-          </div>
-        }
-      />
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "websearch",
-  render(props) {
-    const i18n = useI18n()
-    const query = createMemo(() => {
-      const value = props.input.query
-      if (typeof value !== "string") return ""
-      return value
-    })
-
-    return (
-      <BasicTool
-        {...props}
-        icon="window-cursor"
-        trigger={{
-          title: i18n.t("ui.tool.websearch"),
-          subtitle: query(),
-          subtitleClass: "exa-tool-query",
-        }}
-      >
-        <ExaOutput output={props.output} />
-      </BasicTool>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "enter-worktree",
-  render(props) {
-    const i18n = useI18n()
-    const subtitle = createMemo(() => enterWorktreeSubtitle(props.input, props.metadata, i18n))
-    return (
-      <BasicTool
-        {...props}
-        hideDetails
-        icon="worktree"
-        trigger={{ title: i18n.t("ui.tool.worktree.enter"), subtitle: subtitle() }}
-      />
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "exit-worktree",
-  render(props) {
-    const i18n = useI18n()
-    const subtitle = createMemo(() => exitWorktreeSubtitle(props.metadata, i18n))
-    return (
-      <BasicTool
-        {...props}
-        hideDetails
-        icon="worktree"
-        trigger={{ title: i18n.t("ui.tool.worktree.exit"), subtitle: subtitle() }}
-      />
-    )
-  },
-})
+// Slice 11b.1: read / list / glob / grep / webfetch / websearch /
+// enter-worktree / exit-worktree / todowrite / question / skill renderers
+// extracted to `./message-part-tools-basic`. Side-effect import.
+import "./message-part-tools-basic"
 
 // Render function extracted so both "task" (legacy) and "agent" registrations share one reference.
 const renderAgentToolPart: ToolComponent = (props) => {
@@ -887,155 +677,5 @@ ToolRegistry.register({
   },
 })
 
-ToolRegistry.register({
-  name: "todowrite",
-  render(props) {
-    const i18n = useI18n()
-    const todos = createMemo(() => {
-      const meta = props.metadata?.todos
-      if (Array.isArray(meta)) return meta
-
-      const input = props.input.todos
-      if (Array.isArray(input)) return input
-
-      return []
-    })
-
-    const subtitle = createMemo(() => {
-      const list = todos()
-      if (list.length === 0) return ""
-      return `${list.filter((t: Todo) => t.status === "completed").length}/${list.length}`
-    })
-
-    return (
-      <BasicTool
-        {...props}
-        defaultOpen
-        icon="checklist"
-        trigger={{
-          title: i18n.t("ui.tool.todos"),
-          subtitle: subtitle(),
-        }}
-      >
-        <Show when={todos().length}>
-          <div data-component="todos">
-            <For each={todos()}>
-              {(todo: Todo) => (
-                <div data-slot="message-part-todo-item" data-state={todo.status}>
-                  <Show
-                    when={todo.status === "in_progress"}
-                    fallback={
-                      <Icon
-                        name={todo.status === "completed" ? "circle-check" : "circle"}
-                        style={{ color: "var(--icon-base)", "flex-shrink": "0" }}
-                      />
-                    }
-                  >
-                    <span
-                      style={{
-                        display: "inline-flex",
-                        "align-items": "center",
-                        "justify-content": "center",
-                        width: "16px",
-                        height: "16px",
-                        "flex-shrink": "0",
-                      }}
-                    >
-                      <span
-                        style={{
-                          display: "inline-block",
-                          width: "13px",
-                          height: "13px",
-                          "border-radius": "9999px",
-                          border: "1.5px solid var(--border-weak)",
-                          "border-top-color": "var(--brand-primary)",
-                          animation: "var(--animate-pw-spin)",
-                        }}
-                      />
-                    </span>
-                  </Show>
-                  <span
-                    data-slot="message-part-todo-content"
-                    data-completed={todo.status === "completed" ? "completed" : undefined}
-                  >
-                    {todo.content}
-                  </span>
-                </div>
-              )}
-            </For>
-          </div>
-        </Show>
-      </BasicTool>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "question",
-  render(props) {
-    const i18n = useI18n()
-    const questions = createMemo(() => (props.input.questions ?? []) as QuestionInfo[])
-    const answers = createMemo(() => (props.metadata.answers ?? []) as QuestionAnswer[])
-    const completed = createMemo(() => answers().length > 0)
-
-    const subtitle = createMemo(() => {
-      const count = questions().length
-      if (count === 0) return ""
-      if (completed()) return i18n.t("ui.question.subtitle.answered", { count })
-      return `${count} ${i18n.t(count > 1 ? "ui.common.question.other" : "ui.common.question.one")}`
-    })
-
-    return (
-      <BasicTool
-        {...props}
-        defaultOpen={completed()}
-        icon="bubble-5"
-        trigger={{
-          title: i18n.t("ui.tool.questions"),
-          subtitle: subtitle(),
-        }}
-      >
-        <Show when={completed()}>
-          <div data-component="question-answers">
-            <For each={questions()}>
-              {(q, i) => {
-                const answer = () => answers()[i()] ?? []
-                return (
-                  <div data-slot="question-answer-item">
-                    <div data-slot="question-text">{q.question}</div>
-                    <div data-slot="answer-text">
-                      {answer().length ? answer().join(", ") : i18n.t("ui.question.answer.skipped")}
-                    </div>
-                  </div>
-                )
-              }}
-            </For>
-          </div>
-        </Show>
-      </BasicTool>
-    )
-  },
-})
-
-ToolRegistry.register({
-  name: "skill",
-  render(props) {
-    const i18n = useI18n()
-    const title = createMemo(() => props.input.name || i18n.t("ui.tool.skill"))
-    const running = createMemo(() => props.status === "pending" || props.status === "running")
-
-    const titleContent = () => <TextShimmer text={title()} active={running()} />
-
-    const trigger = () => (
-      <div data-slot="basic-tool-tool-info-structured">
-        <div data-slot="basic-tool-tool-info-main">
-          <span data-slot="basic-tool-tool-title" class="capitalize agent-title">
-            {titleContent()}
-          </span>
-        </div>
-      </div>
-    )
-
-    return <BasicTool icon="brain" status={props.status} trigger={trigger()} hideDetails />
-  },
-})
+// todowrite / question / skill renderers registered via the side-effect
+// import of `./message-part-tools-basic` above.
